+++ data/h2gis/src/main/java/org/orbisgis/data/H2GIS.java
@@ -529,6 +529,20 @@ public class H2GIS extends JdbcDataSource {
     }
 
     @Override
+    public long getRowCount(String tableName) {
+        if(tableName==null || tableName.isEmpty()){
+            LOGGER.error("Unable to get the number of row on empty or null table.");
+            return -1;
+        }
+        try {
+            return JDBCUtilities.getRowCount(getConnection(), TableLocation.parse(tableName, DBTypes.H2GIS));
+        } catch (SQLException e) {
+            LOGGER.error("Unable to get the number of row.");
+            return -1;
+        }
+    }
+
+    @Override
     public String link(Map dataSourceProperties, String sourceTableName, boolean delete) {
         return link(dataSourceProperties, sourceTableName, sourceTableName, delete);
     }