+++ docs/CHANGELOG.md
@@ -6,3 +6,4 @@ eg h2GIS.getSpatialTable("h2gis").columns("land", "st_precisionreducer(st_transf
   .filter("limit 1").getSpatialTable()
 - Add methods on datasource to create, drop indexes, drop table, drop columns
 - Update H2GIS to 2.2.1-SNAPSHOT
+- Add row count method on datasource